flake8 & rename message module
